// src/components/VideConditionalBox/types.ts
export enum VIDE_CONDITIONAL_BOX_ATTR {
  LOW_LIGHT_THRESHOLD = "low-light-threshold",
  HIGH_LIGHT_THRESHOLD = "high-light-threshold",
  MODE = "mode",
  STATUS = "status",
}

export enum VIDE_CONDITIONAL_BOX_MODE {
  LIGHT = "light",
  DARK = "dark",
}

export enum VIDE_CONDITIONAL_BOX_STATUS {
  MOUNTED = "mounted",
  UNMOUNTED = "unmounted",
  MOUNTING = "mounting",
  UNMOUNTING = "unmounting",
}